refactor: Bridge 클래스 제거 및 StateHolder에서 레포지토리 직접 참조
